pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "yolo-llm"
version = "0.1.0"
description = "A hybrid object detection system combining YOLO and Gemini LLM"
readme = "README.md"
requires-python = ">=3.8"
license = "MIT"
authors = [
    { name = "Your Name", email = "your.email@example.com" }
]
dependencies = [
    "numpy>=1.24.0",
    "matplotlib>=3.7.0",
    "Pillow>=10.0.0",
    "ultralytics>=8.0.0",
    "scikit-learn>=1.3.0",
    "google-generativeai>=0.3.0",
]

[project.urls]
Homepage = "https://github.com/yourusername/yolo-llm"
Repository = "https://github.com/yourusername/yolo-llm.git"
Issues = "https://github.com/yourusername/yolo-llm/issues"

[tool.hatch.build.targets.wheel]
packages = ["src/yolo_llm"] 